requirements.txt
ultralytics
opencv-python # used for inferencing
supervision

torch
torchvision

pillow
pathlib
numpy


# Requires chromedriver. Download from https://chromedriver.chromium.org/
tqdm
selenium

scikit-image
requests
bs4

# yolo world
git+https://github.com/openai/CLIP.git


# visualization
matplotlib
streamlit


python-dotenv
tensorflow